requirements.txt
fastapi
uvicorn
openai
qdrant-client
python-dotenv
PyMuPDF
requests
sentence-transformers